requirements.txt
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
decorator==4.4.2
filelock==3.16.1
fsspec==2024.10.0
googletrans==3.1.0a0
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.5
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@90db0de1896c23cbfaf0c58bc2d30665f709f170
pillow==10.4.0
proglog==0.1.10
pydub==0.25.1
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
SpeechRecognition==3.11.0
srt==3.5.3
sympy==1.13.3
tiktoken==0.8.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
